Fix two migrations to satisfy the Migration/AddConcurrentIndex cop

+++ db/migrate/20170207150212_add_indexes_to_mirrors.rb
@@ -11,7 +11,7 @@ class AddIndexesToMirrors < ActiveRecord::Migration
   end
 
   def down
-    remove_index :projects, [:sync_time]
-    remove_index :remote_mirrors, [:sync_time]
+    remove_index :projects, [:sync_time] if index_exists? :projects, [:sync_time]
+    remove_index :remote_mirrors, [:sync_time] if index_exists? :remote_mirrors, [:sync_time]
   end
 end
