+++ db/migrate/20170208144550_add_index_to_mirrors_last_update_at_fields.rb
@@ -11,7 +11,7 @@ class AddIndexToMirrorsLastUpdateAtFields < ActiveRecord::Migration
   end
 
   def down
-    remove_index :projects, :mirror_last_successful_update_at
-    remove_index :remote_mirrors, :last_successful_update_at
+    remove_index :projects, :mirror_last_successful_update_at if index_exists? :projects, :mirror_last_successful_update_at
+    remove_index :remote_mirrors, :last_successful_update_at if index_exists? :remote_mirrors, :last_successful_update_at
   end
 end
